[Refactor] #58 RecommendedContentService 인터페이스 생성
