+++ src/main/java/com/kuit/findyou/domain/information/service/recommended/RecommendedContentService.java
@@ -1,0 +1,10 @@
+package com.kuit.findyou.domain.information.service.recommended;
+
+import com.kuit.findyou.domain.information.dto.ContentType;
+import com.kuit.findyou.domain.information.dto.RecommendedContentResponse;
+
+import java.util.List;
+
+public interface RecommendedContentService {
+    List<RecommendedContentResponse> getContents(ContentType type);
+}
